Refactor: TMDB 인터페이스 공통화 및 구조 개선 - TMDBGenre, TMDBProductionCompany, TMDBProductionCountry, TMDBSpokenLanguage를 공통 인터페이스로 분리
