+++ src/common/interfaces/tmdb-common.interface.ts
@@ -1,0 +1,22 @@
+export interface TMDBGenre {
+  id: number;
+  name: string;
+}
+
+export interface TMDBSpokenLanguage {
+  english_name: string;
+  iso_639_1: string;
+  name: string;
+}
+
+export interface TMDBProductionCompany {
+  id: number;
+  logo_path: string | null;
+  name: string;
+  origin_country: string;
+}
+
+export interface TMDBProductionCountry {
+  iso_3166_1: string;
+  name: string;
+}
