+++ src/genre/interfaces/genre.interface.ts
@@ -1,7 +1,4 @@
-export interface TMDBGenre {
-  id: number;
-  name: string;
-}
+import { TMDBGenre } from '../../common/interfaces/tmdb-common.interface';
 
 export interface TMDBGenreListResponse {
   genres: TMDBGenre[];
