+++ src/movie/interfaces/movie.interface.ts
@@ -1,20 +1,8 @@
-export interface TMDBGenre {
-  id: number;
-  name: string;
-}
-
-export interface TMDBProductionCompany {
-  id: number;
-  logo_path: string | null;
-  name: string;
-  origin_country: string;
-}
-
-export interface TMDBSpokenLanguage {
-  english_name: string;
-  iso_639_1: string;
-  name: string;
-}
+import {
+  TMDBGenre,
+  TMDBSpokenLanguage,
+  TMDBProductionCompany,
+} from '../../common/interfaces/tmdb-common.interface';
 
 export interface TMDBMovieDetailResponse {
   id: number;
